FORTIFY/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fortify-scg"
version = "0.1.0"
description = "SCG construction and pipeline for FORTIFY"
authors = [{name="Research Team"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "networkx>=3.2",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]
